requirements.txt
aiohttp>=3.13.2
colorlog>=6.9.0
homeassistant>=2025.11.0
pip>=21.3.1
ruff>=0.12.2
pytest-homeassistant-custom-component>=0.13.236
pytest-cov>=6.0.0
pre_commit>=4.5.1
apyefa>=1.1.7
